Fix For BUG: When you amend a sale, and click submit and the static page appears and server side validation is executed.  The sale_item_service_landline_business/residential records are deleted, and are only saved when you click commit.  Which means that if you navigate away from the page between submitting and committing, then these details are lost.  I have checked every other table, and landline services are the only tables affected by this bug.

+++ lib/classes/product/type/module/service/Product_Type_Module_Service_Landline.php
@@ -125,19 +125,22 @@ class Product_Type_Module_Service_Landline extends Product_Type_Module
 		else
 		{
 			$product = $products[0];
-			$oldLandlineTypes = array();
-			// Delete the old landline type details
-			if ($product->landlineTypeId == 1) // WIP - Code this properly! 1 = Residential
+			if (!$bolValidateOnly)
 			{
-				$oldLandlineTypes = DO_Sales_SaleItemServiceLandlineResidential::listForSaleItemServiceLandline($product);
-			}
-			else if ($product->landlineTypeId == 2) // WIP - Code this properly! 2 = Business
-			{
-				$oldLandlineTypes = DO_Sales_SaleItemServiceLandlineBusiness::listForSaleItemServiceLandline($product);
-			}
-			foreach ($oldLandlineTypes as $oldLandlineType)
-			{
-				$oldLandlineType->delete();
+				$oldLandlineTypes = array();
+				// Delete the old landline type details
+				if ($product->landlineTypeId == 1) // WIP - Code this properly! 1 = Residential
+				{
+					$oldLandlineTypes = DO_Sales_SaleItemServiceLandlineResidential::listForSaleItemServiceLandline($product);
+				}
+				else if ($product->landlineTypeId == 2) // WIP - Code this properly! 2 = Business
+				{
+					$oldLandlineTypes = DO_Sales_SaleItemServiceLandlineBusiness::listForSaleItemServiceLandline($product);
+				}
+				foreach ($oldLandlineTypes as $oldLandlineType)
+				{
+					$oldLandlineType->delete();
+				}
 			}
 		}
